server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.10.1",
    "django>=5.2.7",
    "django-ninja>=1.4.5",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.38.0",
]
